paypal: include transaction code in unique ID
PayPal's api docs has some language that indicates transaction ID's
may not be unique in some cases.

Their language shouldn't apply to us. It just says that the same
transaction ID may show up in both a balance-affecting and a
non-balance-affecting transaction, and we're only requesting
the balance-affecting ones.

However, for robustness, let's make sure we can't get a clash
anyway, by appending the transaction type code to the ID.