all: added 'gross' and 'net' modes, controllable by 'mode' property
'gross': record the gross amount of the payment in one transaction,
and any processing fees in a separate transaction (so, two transactions
per payment in total).

'net': record net amount of payment (gross - fees) in one transaction
(so, one transaction per payment).

'gross' was the old behavior, and it remains the default.