all: changed default mode from 'gross' to 'net'
'net' is simpler, so it makes more sense as the default.

Which one will work best for you depends on how you record payments
in QBO.

If you include the processing fees as an additional (negative) line
item on the sales receipt, use 'net'.

If you include the processing fees as a separate expenditure, use
'gross'.

Note that some donation gateways like Give Lively add their own
separate amounts to the gross and fee fields to record additional
voluntary charges (e.g., if the user chooses to donate a little to the
gateway, in addition to you). In this case, only 'net' will report
an accurate number, 'gross' will include the voluntary payment to
the gateway inside the gross amount.

See: https://www.givelively.org/resources/learn-how-stripe-processes-and-displays-voluntary-contributions